+++ scripts/install.sh
@@ -1,4 +1,5 @@
+#!/bin/sh
 # install poetry and the dev-dependencies of the project
 python -m pip install poetry
 python -m poetry update --lock
-python -m poetry install -E redis -E rabbit -E kafka -E azure
+python -m poetry install -E all