+++ scripts/unittest.sh
@@ -1,2 +1,3 @@
+#!/bin/sh
 # run the unittests with branch coverage
-python -m poetry run python -m pytest --cov-branch --cov=./yellowbox --cov-report=xml tests/
+poetry run pytest -n auto --dist loadfile --cov-branch --cov=./yellowbox --cov-report=xml tests/